refactor: remove old unused code
