+++ src/lib/addTrailingNewLine.ts
