+++ src/lib/buildFieldProperties.ts
