+++ src/lib/buildPropertyName.ts
