+++ src/lib/collectCustomTypeFields.ts
