--- src/lib/getAPIIDPath.ts
@@ -1,45 +1,0 @@
-import type {
-	CustomTypeModel,
-	CustomTypeModelField,
-	CustomTypeModelSlice,
-	SharedSliceModel,
-} from "@prismicio/client";
-
-import { PathElement } from "../types";
-
-import { isCustomTypeModelField } from "./isCustomTypeModelField";
-import { isCustomTypeModelSlice } from "./isCustomTypeModelSlice";
-import { isSharedSliceModel } from "./isSharedSliceModel";
-
-type GetAPIIDPathConfig = {
-	path: [
-		PathElement<CustomTypeModel | SharedSliceModel>,
-		...PathElement<CustomTypeModelField | CustomTypeModelSlice>[],
-	];
-};
-
-export const getAPIIDPath = (config: GetAPIIDPathConfig): string => {
-	return config.path
-		.map((element, i) => {
-			if (
-				isCustomTypeModelField(element.model) &&
-				(element.model.type === "Group" || element.model.type === "Slices")
-			) {
-				return `${element.id}[]`;
-			} else {
-				if (element.id === "items") {
-					const previousElement = config.path[i - 1];
-
-					if (
-						isCustomTypeModelSlice(previousElement.model) ||
-						isSharedSliceModel(previousElement.model)
-					) {
-						return `${element.id}[]`;
-					}
-				}
-
-				return element.id;
-			}
-		})
-		.join(".");
-};
