--- src/lib/getHumanReadableFieldType.ts
@@ -1,55 +1,0 @@
-import type { CustomTypeModelField } from "@prismicio/client";
-
-type GetFieldHumanReadableTypeConfig = {
-	model: CustomTypeModelField;
-};
-
-export const getHumanReadableFieldType = (
-	config: GetFieldHumanReadableTypeConfig,
-): string => {
-	switch (config.model.type) {
-		case "StructuredText": {
-			const isTitleField =
-				config.model.config &&
-				"single" in config.model.config &&
-				config.model.config.single &&
-				config.model.config.single
-					.split(",")
-					.every((blockType) => /heading/.test(blockType));
-
-			return isTitleField ? "Title" : "Rich Text";
-		}
-
-		case "IntegrationFields": {
-			const catalog = config.model.config?.catalog;
-
-			return `Integration Fields (Catalog: ${
-				catalog ? `\`${catalog}\`` : "*unknown*"
-			})`;
-		}
-
-		case "Link": {
-			switch (config.model.config?.select) {
-				case "document": {
-					return "Content Relationship";
-				}
-
-				case "media": {
-					return "Link to Media";
-				}
-
-				default: {
-					return "Link";
-				}
-			}
-		}
-
-		case "Slices": {
-			return "Slice Zone";
-		}
-
-		default: {
-			return config.model.type;
-		}
-	}
-};
