--- src/lib/isCustomTypeModelSlice.ts
@@ -1,14 +1,0 @@
-import type { CustomTypeModelSlice } from "@prismicio/client";
-
-import { hasOwnProperty } from "./hasOwnProperty";
-
-export const isCustomTypeModelSlice = (
-	model: unknown,
-): model is CustomTypeModelSlice => {
-	return (
-		typeof model === "object" &&
-		model !== null &&
-		hasOwnProperty(model, "type") &&
-		model.type === "Slice"
-	);
-};
