--- src/lib2/buildTypeName.ts
@@ -1,11 +1,0 @@
-import { pascalCase } from "./pascalCase";
-
-export const buildTypeName = (...parts: string[]) => {
-	let name = pascalCase(parts.filter(Boolean).join(" "));
-
-	if (/^[0-9]/.test(name) ? "_" : "") {
-		name = `_${name}`;
-	}
-
-	return name;
-};
