--- src/lib2/pascalCase.ts
@@ -1,9 +1,0 @@
-import {
-	pascalCase as pascalCaseBase,
-	pascalCaseTransformMerge,
-} from "pascal-case";
-
-export const pascalCase = (input: string): string =>
-	pascalCaseBase(input, {
-		transform: pascalCaseTransformMerge,
-	});
